Yet another O(n^2) bottleneck. This time in exportTiles
By using a treeSet instead of a linked list gives a huge performance boost
on large midlets. This is currently bit of a hack, as rather than propagating
the TreeSet through out and replacing the linked list alltogeter, this only
creates the TreeSet locally were it becomes a bottleneck. This was the easiest
to implement though.