Use the standard java xml parser instead of MinML2
Not sure what the reason is to use MinML in the first place,
but trying to use the standard XML parser instead. This change
is very painless, as both use the same interface.

The reason for this change is, that the most time is used
up in the xml parser now. See if this improves anything.
Looks like it may save a few percent, nothing dramatic though