Made tags private in Entity and added getters and setters
Currently all Entities had a Hashtable created for potential tags.
As many of the nodes may not have tags, this seems wasteful in memory
consumption. So only allocate a HashMap if there are actually tags in
this entity. In order to make it easier to experiment with the data
structure, abstract the tags variable behind getters and setters of the Entity