improved GPX handling, don't buffer everything bevor transmiting
There is a possibility to transmit the length of a file in the obex headers
I though this might be necessary. In order to calculate the length, one has to
buffer the entire gpx log. This can easily exhaust all available memory. Revert this change
and construct the gpx on the fly, not sending the length header