BugFix: dropCaches could cause Null pointer excpetions
we were not catching all of the cases where null pointers could appear
due to the call of the dropCache in a low memory condition.

Some of these could be fixed with if != null checks. Some required additions
of synchronized to guard against thread safty issues. I hope this doesn't effect