Added pre-processor directives to disable OBEX support
Unfortunately many Nokia phones have broken implementations
of JSR-82. They don't implement Obex support. So we need to make
this option compile time disablable. Unfortunately j2mepolish also
lakes in this respect, as it does not offer a possibility to detect
which devicses support obex and which don't. At the moment, I have
just invented my own pre-processor directive to guard the obex support.
So if one wants to enable it, one has to add this to the devices.xml file