BugFix: SearchNames Thread could crash, as type 0 could mess up termination
We were using a type of 0 to indicate termination of a loop in the map file format.
Unfortunately however type 0 has become a valid type with the migration to the
stylesheets. This case premature termination of reading an entry, causing data
corruption for the following entries.

Bumped Map file format to 7