Fixed some forgotten float to short coordinate conversions in SingTile.walk()
SingeTile.walk() was added to HEAD after the Optimisations branch was forked off
where I did the original conversion to 16bit. I overlooked it when merging back
Optimisations branch to HEAD. This function seems to be used for routing, but I
am not quite sure what it does, so I haven't really tested this change. The previous
state definately looked wrong though.