BugFix: When ways are split in OxParser, tags weren't correctly duplicated
When a node referenced by a way was not in the list of nodes (e.g. outside the bounding box)
the way got split at that node. The current tags got duplicated for both ways. However,
at the time of split, not all the tags of the way have necessarily been read. So
keep track of the duplicate ways and ensure they all have the same tag.

This is part of bug #1952516. I haven't verified that this is covers all cases yet