Feature: Respect the osm layer tag
We now split the rendering order according to the layer tag.

We first render all Areas, the lowest layers first, then higher layers. Once all
areas have been drawn, we then draw the ways. Again according to the layer tag.

In order not wast to much resources, it is currently limited to -2 .. +2

The layer information was put into the SinglTile rather than the way it self,
in order to save some memory. The Way class is already the biggest memory consumer,
as we have large amounts of Ways in memory at a time.