Feature: Make debugging log severity a runtime option
It is now possible to set at runtime which debug severity messages
one wants to see in the logs. Error and Fatal can not be turned off,
as these should be critical enough to always display.

The debug messages are still guarded by pre-processor directives,
so one can only enable them if the compile time debug severity is
set accordingly.