Determine the maximum memory the jvm can support
Some jvms, such as the ones on Sony Ericsson phones, have a dynamic heap allocation
and therefore the call to Runtime.totalMemory() might not reflect the true capabilities
of the device. We however use this call to figure out if we still have resources available,
or if we need to be more agressive in cleaning up caches to try and not hit the case of
OutOfMemory errors. This ends up being overly conservative on these devices though. So
instead of using the reported Runtime.totalMemory, use our own estimate of the available heap.

This is determined, by allocating as much memory as we can until we hit an out of memory error,
at that point the jvm should have increased the heap to the maximum possible, at which point
the totalMemory should be mostly accurate.