Get the JAD/manifest information from the midlet and don't recreate it in Osm2GpsMid
While repacking the midlet coming from the GpsMid compilation and adding the map data,
Osm2GpsMid used to overwrite the manifest file and JAD file with a hardcoded version.
Not only does that unnecessarily dupplicate data, this can actually cause real problems,
if these data change due to changes in the GpsMid compilation. I hit this problem,
when trying to debug some obfuscator/proguard/j2mepolish issues. Depending on the settings
of the obfuscator, the main class GpsMid could either be in its original directory of
de.ueller.midlet.gps or it could be collapsed into the root directory. The compilation
process of GpsMid correclty changes the manifest and JAD files, but we then overwrote
the correct version with incorrect hardcoded information from Osm2GpsMid.

With this commit, we no longer write a manifest file in Osm2GpsMid and simply use
the one comming from GpsMid. The JAD file is derived from the manifest file with a
few additional lines e.g. jar-size.