Increase the size of the Name-XXX.dat files
Previously, each file only had about 100 Names in it. Given that the names were
quite well compressed, this caused the average names file to be about 1 - 2kb.
This ends up creating unnecessarily many files and causes many warnings if using
filesystem maps.

Instead split the name files after a specific size of the file. The size is
the same as the maxTileSize parameter in the properties file (by default 20kb)