Optimisation: Move the setEntity call out of the inner most loop
the setEntity expects an array of nodes as floats. Given that we only have the
relative coordinates, we have to convert each coordinate. As the getFloatNodes
call converts every coordinate in a SingleTile makeing it reasonably expensive,
especially if we were calling it in the inner most loop of the paint()
method. Move it at least one level up. This is still not ideal and shouldn't
really be called at all in the paint method, but that would have needed some
changes to the routing side of things, so I will leave it for later.