BugFix: 16bit overflow in calculating the screen center relative to tiles
When zoomed out very far, the relative scales of the tiles where small compared
to the distances on the screen. This caused nummerical overflow when converting
the screen bounding box to tile relative coordinates. If this happens,
simply set the coordinates the the 16bit maximum, as the exact values
shouldn't matter