Implement a median search for float arrays
This is currently done in a insanly inefficient way using bubble sort rather
than a proper O(n) algorithm. This function will be part of a bug fix for
the GPX tile splitting code. N for the moment should therefore not be bigger
than 90, so hopefully will work till this function gets recoded.