BugFix: GPX recording crashed when recording longer stationary tracks
The GpxTile code could get into an infinite loop of trying to split
up GpxTiles. If more than 90 trackpoints are stored in a Tile, it is
split in a kD-Tree fashion. We selected the coordinate at random
at which to split the larger dimension. However if more than 90
trackpoints have identical coordinates, a split did not reduce the number
of points in a tile and we could thus get into a infinite loop.

We now use a median finding routine to split more sensibly. Furhtermore
we check to make sure that either subtile had at least some of the points