Improved Relations parsing
Slightly improve the support for Relations parsing. It now filters
relations according to if the member ways and nodes are in the bounding
box. This needed some changes to the datastructure holding ways to search
for ways by id.

The relations support still doesn't actually do anything yet though, and for
the moment I don't really have concrete plans to support any. This is more
about seeing how many relations we currently ignore