Patch: [ 2044193 ] Increased robustness of track and waypoint storage
This is a patch by Markus Burle

Rather than keeping the recordstore open all the time, which seems to be prone
to corruption if the midlet crashes, only open it for a short time while it is
needed. Apparantly it was possible before to loose all of the stored tracklogs
due to corruption which is very bad!