Remove the temp dir again after we have finished the conversion
So far we have left behind one temp dir per conversion process, which is not very
nice. With this commit, by default we will remove the dir.

As there are some reasons for keeping the directory. e.g. if you want to use the map
data, it is possible to specify this in the .properties file with  keepTemporaryFiles=true