Allow to specify bounds on the command line of Osm2GpsMid
With this patch one now no longer needs a .property file either,
but can specify the bounds on the command line and take otherwise
the default parameters

the syntax is e.g. java -jar Osm2GpsMid.jar --bounds=52.0,1.0,50.0,-0.1 england.osm.bz2

With this it should now be possible to create a midlet with no other
file needed than the Osm2GpsMid jar.