Add a new datastructure LongArrayMap to replace our usage of HashMap
Using HashMaps to store the Node and Way lists resulted in quite a lot
of overhead in memory usage. The memory used in HashMapEntries and hash keys
was more than the memory for the Nodes and Ways itself. With the LongArrayMap,
we can reduce the memory overhead by quite a bit by storing it in large arrays.

We exploit the fact that we add entries in a sorted mannor (.osm files
contain Nodes and Ways in increasing id order), so adding it to an array
is easy as we can simply append it at the end.

Currently this is not a complete implementation of the Map interface but only
of those parts currently needed by Osm2GpsMid.

With this commit, the datastructure is not yet actually used.