BugFix: tag maxspeed= was not correctly handled (double conversion to m/s)
Due to the double conversion, explicit maxspeed= ways were given a lower maxspeed
than they should have had.

At the same time this improves the special case handling.
So far we didn't deal with fractions in the maxspeed tag as happens when converting
from mph to km/h.
If the tag contains the text km/h or mph, trim the characters and if necessary convert to km/h