BugFix: maxLon of the projection was calculated wrongly for -ve longitudes
The maxLon bounding box was calculated incorrectly for -ve logitudes to be
approximately 0 rather than the actual negative longitude. This caused
the bounding box of projections to the west of the meridian to extend up to
the meridian and therefore far too much was claimed to be on screen causing
a masive slow down in rendering speed.

The problem is that Float.MIN_VALUE is the minimum positive value
rather than the maximum negative value and as such "if (n.radlon > maxLon)"
was never true for negative longitudes.