DataStrucutre: Added SmallArrayMap, a very compact Map interface implementation
This Datastructure is optimized to have as little as possible memory overhead.
It stores both keys and values in the same array, so the only overhead is
The object header of the SmallArrayMap and of the one key_value array plus of cause
one pointer per key and value.

To search the map, a linear search is performed, so this is only usefull if the maps
are very small. It was developed for the tag maps of entity.java for which there can
be millions, each only very small. This saves some memory compared to the HashMap we
were using before.

This commit does not activate this feature yet.