Feature: A first stab at adding exif information to a picture containing GPS coordinates
It is useful if the picture immediately contains the GPS coordinates in its exif data,
rather than having to use a seperate geotagging program to match the timestampt of
the image to a GPX file.

With this commit we attempt to add the Lat/Lon and altitude information.

It is still preliminary and rather messy, as we don't fully handle exif, but simply
copy in a binary blob into the image and patch up the Lat/Lon values. This means,
that if the image already contains an exif block, we end up having two blocks in
the file, as GpsMid copies its block in front, without touching the old exif block.

Many thanks go to Alessio Zanol for figuring out how to easily add exif to images and
providing the instructions of how to do this