Attempt to fix the hideNotify bug by removing setTitle(null)
Sony Ericsson phones were having troubles with occasional map freezes after which
the map no longer updated although the program continued to run just fine otherwise.
The problem was that the JVM thought the midlet was in the background and no longer
called the paint() method, even after an explicit repaint().

It looked like this behaviour strongly corrolated with the call to setTitle(null). So
this patch removes the use of setTitle and replaces it with our own implementation of a title bar.

Hopefully this will fix the map freeze bug, but that will mostly tell once more people have tested
this on a variety of phones.