BugFix: Application could crash with null-pointer-exception when panning during route calculation
If map screen was disabled during route calculation, the application could crash if one attempts
to pan the screen. This was caused by the fact that route calculation suspended the map drawing via
stopImageCollector() which sets the image collector to null. Panning the map screen however unconditionally
accessed the imageCollector causing a null pointer exception in the UI thread crashing the entire application.