Separate storing of a GPX tracklog from closing the track recording
This change allows to store intermediate versions of a tracklog and continue to use it afterwards.
I.e. it is now possible to have "safty copies" so that we don't loose the entire tracklog if the phone crashes.

For the moment, we only manually trigger this when suspending the tracklog