BugFix intTree would fail if elements are added in reverse order
The problem was our use of the sign bit as an indicator to say if
we should add or overwrite an entry. The problem was with 0 not
being signed. So when we were trying to add an entry at position 0
we would simply overwrite that position rather than add it.