This fixes bug #2183168:
When importing gpx files now always UTF-8 encoding is assumed regardless of any given encoding parameter in the gpx file.

This change is necessary to be able to parse the encoding of GpsMid's exported gpx files again after fixing bug #2164148.

The remaining problem that the encoding parameter in the gpx file is ignored is described in #2208230