BugFix: The map screen was not correctly responding to the camera cover open key event anymore
Since the move to the more generic key mappings, the camera cover on a K550 no longer opened
the Camera Dialog. The problem was, that this special key does not report a key released event,
as it is a switch and not a proper button. Therefore we need to include a special case treatment
in the keypressed routine