BugFix: Show the correct error dialog when the map version is not correct
closes [ 1941239 ] Exception "Wrong map file format" hides alert

We were by now only getting a null pointer exception if the map version was incorrect.
The log file still contained the information about the version conflict, but this information
was never shown on screen so it might easily haven been missed.

It is also now no longer possible to continue if the map version disagrees