BugFix: Speed limits were not correctly shown if not specified in pure integers
We had previously fixed up the maxSpeed calculation in routing to deal with different units and do a proper conversion,
however the displaying of the speed limit did not use this conversion code but directly tried to use the tag.

So this refactors the code for both to use the same conversion.