EditingSupport: Store the osm ID for every way and poi node, to be able to link the GpsMid data to the real OSM data for potential editing
If we enable support for editing in the .properties file, then the map data will now contain the actual openstreetmap id. This allows to
retrieve the full data from the OSM servers and in future at some point maybe editing too.

This increases the size of the map data quite considerably, so this feature should only be enabled if you really need it.

As it changes the map format, increase the map version number to 20. This needs the equivalent changes in GpsMid