EdditingSupport: Add a possibility to display the full openstreetmap xml data of a selected way
With this feature, you can display the full openstreetmap XML for the current way. This required internet connectivity,
as the data is not stored in the map data, but has to be retrieved directly from the openstreetmap servers.

As this adds some overhead both storage wise and processing wise, this feature can be disabled both at runtime in Osm2GpsMid
or by compile time by disabling the pre-processor directive EDITS_ENABLED.

The pre-processor directive is disabled by default, so at the moment there shouldn't be much overhead unless you actively enable this
feature during compile.

This increases the map version to 20 and requires the equivalent Osm2GpsMid commits.