Add manual rotation.
When in manual rotation mode you can rotate the map using left/right or 1,4,7,3,6,9,8. 2 sets the map to North. Any other key ends the manual rotation mode.

This commit also experimentially sets the SAVE_WAYPOINT command to the short press of 9. Thus the manual rotation mode can be started with the long press of 5.