Patch "Make BACK key work in GuiWaypointSave - ID: 2489058" by mbaeurle:
According to the J2ME specification (see e.g. in this Javadoc in the
paragraph "Type":
http://java.sun.com/javame/reference/apis/jsr118/javax/microedition/lcdui/C
ommand.html), Commands that represent a BACK type action should be declared accordingly.
This was not done in GuiWaypointSave.java while it was done in other places
such as the tracks or waypoints lists.
Obviously, I had copied and pasted it to GuiWaypointEnter.java as well.
:-|
This bug was the reason why the BACK key was not working (on phones that have it, naturally) in these dialogs.

This patch fixed this, it simply replaces Command.OK with Command.BACK.