Reverting last commit of sk750 to help in merge conflicts
I had some time on the plane yesterday, so I have begun to refactor some code to make keyboard short cuts configurable.

Unfortunately as I hadn't mentioned it to anyone, these now conflict with the changes sk750 did at the same time. I think
it is easier if I revert this commit, then commit my changes and the try and fix the issue of this patch in a new way.

In case this causes more headaches for you, as you had other uncommited changes too, then feel free to revert everything back again
and we will have to see how to go about this.

Sorry for this.