qwert keyboard search support in gpsmid
This is a patch by jkpl

This uses NumberCanon to map alphabetic characters to numbers so that the
search can be used with qwerty keyboards too. This is not perfect, as there
are fals positives in the search, as several letters get mapped to the same
numbers, but better than nothing.