Move the default key configuration into a keyMap.txt file
This is another step towards making the keycode configurable. Rather than hardcodeing the default
key configuration, with this patch, the config is loaded from the keyMap.txt file in the bundle.

The format is currenly not particularly user friendly, as it requires to know both the keycode and the internal
index of the action to which the key maps. Each line in the file corresponds to a single mapping. The first number
represents the keycode, and the second number represents the index into the action array.