Overview/Filter Map: Add "containing..." as an option to Name Check. When you select this option you are offered to enter a name part  that must be contained in an element's (POI's, way's or area's) name to be displayed. This is hopefully useful during mapping to display e.g. only POIs containing "FIXME". On the other hand it's also amazing when using the map: You can e.g. display only supermarkets with a certain name when you are hunting for a special offer that is already sold out in your local market.
Please note:
a) If no elements (POIs / areas / ways) are selected, the name check will be performed on ALL elements of the selected type.
Thus in most cases it will be enough to only enter the string to be contained  to get the desired result.

b) If you also select one or more elements from the list, the name check will be performed only on the selected elements and the other ones will get filtered out.