Render a bit more than only the screen for smoother scrolling
If we render a boarder around the screen, it is possible to display a full screen
during panning again before it has rerendered, giving a smoother feeling. This might
cover elements of feature request #2504679.

As we are now rendering a significantly larger area than before, this will have an impact
on the speed each individual frame gets rendered at, so it must be carefully balanced between how much extra we render
and how fast one can scroll without seeing an incomplete rendering.

This used to work previously, but got broken at some point in the past.