BugFix: Hopefully get the JSR-172 implementation of the XML parser working again
The implementation it self worked previously, but the obfuscation step broke it, so it
didn't work on anything else than the test builds.

The problem is that MinML, which we include, implements some of the same classes and methods
that are interface code in the JSR-172 implementation. However, as the classes were included
due to MinML, the obfuscator thought it can rename them, as they aren't libraries and therefore
has bother caller and callee available for renaming. Of cause that is not true for the jsr-172 implementation,
causing the code to use the base method, rather than the overwritten method of startElement.

By adding a keep parameter to the build.xml, this partly fixes it. However there might still be
problems with there being two implementations of the same class in the jvm and indeed the new
version of proguard (4.3) won't even compile as it is complaning about this issue. The real fix I believe
would be to rename the classes in MinML to not clash with the JSR-172 classes.