Warn if logging is enabled but not compiled in
People have tried to enable logging in the Debug part of the setup, however
logging by default is no compiled in for performance reasons. Previously
there was no indication of the compile time issue, other than that the logs
stayed empty.

With this commit, we now warn the user that logging is not compiled in when
they try and enable it.