Feature: Use the direction of travel to estimate the way we are currently travelling on
So far we have only used the distance between the current GPS position and a Way to determine
which way we are on. This however causes the way information often to be wrong when travelling
on a road with many side streets, as as soon as the GPS position is slightly off the road, the
side streets will be closer than the main road while crossing them at right angles. By taking the
direction of travel into account, we can penalise the roads at 90 degree, such that we still think we are
on the main road.

The algorithm calculates the dot product between the current direction of travel and the direction of the
road. As an approximation, the dot product is used as a linear penalty on the distance, with a right angle
roughly corresponding to a distance offset of 100m. The direction calculations are performed on the projected
ways. This should more or less work for 2D projections, but won't work anymore, should we ever get a 3D projection
mode.

As there are a bunch of approximations involved to keep the computational requirements as low as possible,
the heuristic will most likely need a lot more tweaking, once we have some real life experience.