BugFix: [ 2545019 ] Negative speed limit on highways
This I think might fix it. The problem is that we were using bytes, which are 8bit signed, so can only go up to 127. I have changed it to using
int in the program. The data in the file is still 8bit, but is meant to be unsigned. As I didn't have a way handy with a 130 speed limit, I haven't
actually tested it yet though.