Enable the ability to directly upload GPX traces to your OpenStreetMap account
Previously we were lacking the ability to enter your account details, so it was
not possible to upload. With this commit, it should now be possible to do so.

In order to upload a GPX tracklog to your OsmAccount, you need to set your
user name and password in the OSM account setup dialogue. Then you have to select
Upload to OSM in the GPX receiver setup. Now when ever you upload a GPX, it will be uploaded
to the OSM server.