Add support to store and reload Configuration settings to an external file
This is the first step towards implementing feature request 2491888, but doesn't
contain the UI yet to use it.

This feature is particularly useful on Nokias, as they don't reuse record stores
when reuploading the midlet. Therefore you have to reenter the entire configuration
everytime you want to update the map data.