Editing: Improved the way editing capabilities
This commit adds a primitive way properties presets menu
Adding even a single key value pair on the numeric keyboard of a mobile
is painful, so try and avoid this by having some presets. Currently the set
of available presets is hardcoded to the ones I have needed so far. This will need
a more flexible way.

Add the ability to see in which direction a way is pointing. The information is taken
from the way information in GpsMid and not from the XML retrieved from the server.
I hope this doesn't cause more damage than good, but if you want to add a oneway tag
it is necessary to see in which direction the way goes.