[ patch #2624059] Multiple LocationMsgReceivers for one LocationMsgProducer
This is an adaption of the patch by Jan Rose (black_ivory)

This patch add the possiblity to have more than one
LocationMsgReceiver for one LocationMsgProducer so the data is sent
directly to the WaypointGui or any other part where it's needed
instead of piping it through the Trace class.  which only keeps the
last Location. I think this is useful for many other parts as well and
could help to make the Trace class more focused in it's functionality.