[ Feature Request #2424559 ] Area POIs
due to the importance of this feature request, I have decided that a crude hack is probably better than
no solution, so this will hopefully do until we have something better.

For the moment, this patch simply checks all ways against the POI legend and if it matches one of the POI types,
the first node of the way gets set to be of type of the detected POI.

This solution is somewhat ugly, as the POI image gets placed in some corner of the Area
and the name of the POI and the area overlap, but both of these can probably be somewhat
fixed. A proper solution is however still required eventually. On the test midlet of
central London, this patch increases the number of POIs by about 10%, but this figure will
greatly vary from area to area.

[# Your branch is ahead of 'origin/master' by 153 commits.