[ FR #2720566] (not quite) One Touch Audio Recording
This isn't fully implementing the feature request, but gets us a little closer

It is now possible to put the audio recording on a key shortcut. At the moment
it is a double press on #, but probably needs a better key in he future.

So to start recording you double press # and to stop it you again press double #