Bugfix: Increase the number of open alerts now only if a timer for the previous alert only if a display timeOut timer for the previous one already has been scheduled.
It turned out during adding a display timeout timer for title messages,
that if an extra message came in, while the timer had not been started in the paint code, the number of open messages was increased but only one timer decreasing the number of open messages on timeOut was scheduled. Thus because the number of open messages never decreased to zero again the title message never disappeared again when messages were received faster than timers were established in the paint code.

The same applies for the alerts.