CellIDLogging: If no data was written to the cellid log delete it again.
Instead of creating the log file and then deleting it again, it would have been
better to not create it in the first place. However, that didn't work,
as it would have not been possible to say if it will work later on,
or if the user for example will reject the security dialog. The error
propagation paths therefore didn't work out.