Add a persistent cache to the cell-id location provider
Previously we had to retrieve all cell-ids from OpenCellId.org via the internet.
This is not ideal for a client that tries to be offline as much as possible.

Although we still need to retrieve the the information over the internet the first
time, at least we now persistently cache it in a record store. As there is no method
of deleting the cache at the moment, we only store entries that returned a valid
location, and not those that don't yet have an entry in OpenCellId.org to be able
to catch newly added data.

Eventually we will want to be able to populate this list offline (e.g. include
all cell-id positions during Osm2GpsMid), but that is further down on the list
of things to do.