Take direction penalty also into account for the distance to the closest segment on route line path.
This hopefully improves the handling when going below a bridge and then onto the bridge so you don't get the wrong instruction.

I had no opportunity again to give this a try, but at least everything else still seems to work fine with this change.

This commit also combines the nearest and the actual routable way as we now  always use direction penalty when fast enough.