Reset the persistent cell-id cache if anything goes wrong with it
I was not able to use the cell-id provider anymore, as my persistent cache got corrupted.
Instead of silently failing, it now simply deletes the cache and starts fresh whenever it detects
problems with it. This should be fine, as it is only a cache and the only problem is that one
needs to get fresh data from OpenCellId.org

fine, as it is only