Update editing support for API 0.6
Changed the API-url to the test server mysql.dev.openstreetmap.org

This is the only server that currently has at least some real data in it
(it has an old planet exctract of GB), so that one can use it together
with the generated data from Osm2GpsMid.

If you want to test it, you need to use the microemulator rather than the sun emulator,
or use your mobile if you are fine with paying for internet. Then enable editing
in your properties file. You will also need a new account on mysql.dev.

Otherwise, it should have the same (limited) editing support as before, just now
ready for API 0.6! :-)

There is still a long way to go to have good editing support, but it is hopefully a decent
and already useful start.