Update ROMA and xapi links to 0.6
Xapi doesn't currently seem to work, so this isn't tested.

the ROMA link is also no longer purley ROMA, but a combination
of ROMA and TRAPI. The new link seems to work and deliver data,
but the data being returned is not identical to that of the main
API. e.g. some tags (such as tiger and created_by) are removed. Never
the less, it seems to be working with GpsMid