Limited support for using cellids in S603rdFP2 devices
Fron FP2 onwards S60 supports retrieving cellids via System.properties
Unfortunately Nokia does not allow to retrieve the LAC (Location Area Code)
of the cell, which is needed to uniquely identify it. Although not ideal,
this is sufficient to retrieve existing Cells from OpenCellID.

So this patch adds support for S60Fp2 devices as a location provider,
but not as a recorder of CellIDs, as if you want to submit these to
the DB, they require all information including the lac.