[GpsMid Patches 2787770] from anonymous: Fixed JSR179-Problem on Samsung SGH-G810
Original description:
On Samsung SGH-G810 I got an exception when using the internal gps. I found
out that the JSR179-Implementation throws an NullPointerException when the
specified criteria for LocationProvider are not supported. After catching
this exception the internal gps worked well.
Additional I changed the way checking supported criteria. The Samsung
supports only altitude but not speed and course. With the old
implementation it became a LocationProvider without special criteria. The
new implementation guarantees the best criteria for all phones.