[ gpsmid-Bugs-2781642 ] Routable must be a way flag instead of way description flag
With this commits (GpsMid and Osm2GpsMid) routable becomes a way flag. This flag is realised as a byte. Later on it is possible to use this byte to indicate for up to eight route modes (motorcar, bicycle, etc.) if the way is accessible. The commit prepares also up to 8 different typical speeds for a WayDescription. Currently only bit 0 of the Way routeModes / WayDescription typicalSpeed[0] gets used.

This increases the map format version to 26.