BugFix: Close the log file on exit
We weren't closing the log file on exit. Without this, not all of the log was written
on some devices and thus the log file could be empty despite there having been
messages to log.

This also means that presumably in the case of a crash, the log will also be empty,
as in that case we don't get to close the file cleanly.