We somewhere have a null pointer exception in the Editing code which crashes GpsMid completely
I haven't figured out where exactly this comes from so I have put some extra null checks in to
try and help. I don't think it is fixed though, as I have hit it once again. It felt less frequent,
but it might be that it has something to do with network failures or delays that are intermittent.