Show if we are successfully logging cells
The cell logging was previously shown as soon as we have a valid method for logging,
but did not tell if currently we have all necessary data. E.g. we might not be currently
connected to a cell as we are in a tunnel. It now shows in red if it is attempting
to log but currently can't

because