Feature: Allow for offline positioning with CellID on those phones that support CellID
If the data is included by Osm2GpsMid in the map file, this will now automatically be used.

The order of preference of retrieval of the different methods is now. Memory cache (gets reset every time you start GpsMid),
offline map cache (Gets updated once you recompile GpsMid with Osm2GpsMid), Persistent cache, Online retrieval from OpenCellID

So if the data was not included in Osm2GpsMid, it will still fallback on retrieving online cell info, but one can prevent that
by dissallowing GpsMid to connect to the internet.