[ gpsmid-Patches-2517632 ] Ask for track name on start/stop recording Submitted By: jkpj (jkpj) Summary: Ask for track name on start/stop recording
Initial Comment:
Here's a patch to add functionality to make GpsMid
ask for GPX track name on start and/or stop of GPX recording.  The behaviour (whether to ask for name) is configurable in Setup, at Recording Rules.

The idea is to make it easy to keep track of which track is which.  With both switches on, GpsMid will ask for track name (prompting the user with the default timestamp name) when starting and stopping recording.  When the user enters something like vehicle/location for start and location at end, it'll be easy to see from track manager and/or GPX file names which tracks are from where.