Create a new pre-processor directive to in future guard all online activity.
The plan is it to have different midlets for when you want to be able to use online features
and when you don't. This should give a better safeguard against accidentally going online and
incuring charges.

The online pre-processor directive was added to the Generic-editing device. But some renaming
and combining the two is probably useful.