Add online links to the weatherforcast and wikipedia articles for the current location
This patch adds a menu option to go to websites relevant to the current location.
At the moment this includes looking up close by wikipedia articles and the current weather forcast.

The weather forcast is thanks to http://www.wunderground.com/
Wikipedia articles are retrieved via the webservices of geonames.org

The wikipedia service of geonames is unfortunately not quite ideal yet. It goes via an RSS feed
and doesn't link to the mobile optimised pages of wikipedia.