APIfix: The upgrade of the OSM servers to Rails 2.3.2 broke GpsMid
It looks like the latest rails no longer supports the _method=put parameter that GpsMid needs
due to the broken HTTP stack of j2me.

Instead we now use the HTTP_X_HTTP_METHOD_OVERRIDE header, as suggested in
https://rails.lighthouseapp.com/projects/8994/tickets/2289-_methodput-ignored-for-xhr-and-xml-requests

This means that all previous versions of GpsMid are now broken with respect to editing, and it is
necessary to update to at least this version to be able to edit again.