Prepare turn restriction object model for via ways.
We actually put turn restrictions based on via node and via way refs into the same structure and flag turn restrictions with via ways. This must be resolved later on before or during turn restrictions are calculated, that is when all ways are read in and all route nodes have been created. For now we just ignore the turn restrictions with via ways during turn restriction calculation.