Disable On-Screen-Keypad for Nokia S60 touch devices
On S60 touch devices, by default about half of the screen estate
is used up by huge onscreen keypad. This keypad however only emulates
the cursor keys and a few game action keys. The function of these (mainly draging the map) are implemented
as proper touch screen controls, so there is no need to waste
screen estate with the virtual keypad.

GpsMid is still not perfectly usable completely without a keys, but the ones
available through this emulation are just not the ones we need.