[Editing] Use a preset menu to more easily select a POI type to add
Adding a POI now works in the following way:

You select "Retrieve Node" (needs to be renamed to something more sensible) in the main map.
This will display a list of POIs from which you can choose. After confirming the type,
it gives you a list of tags pre-filled in with the values from your selected POI type.
These tags can now be freely edited and then uploaded to the OSM database.

The current cursor position will be taken as the place to add the POI later.

It is important to notice, that no POI will appear in GpsMid, but it still is in
the OSM database, so please don't upload it twice.

More improvements are still necessary