[GpsMid Patches #2813596] Updated painting of waysegments to draw them connected - by pos_ei_don_
From the tracker item:
"Attached is the patch and a screenshot which changes/corrects the painting
of waysegments. I used an early version of this code already in the
paint-bridge patch but I found out that this should be corrected in
general, not only for bridges. As it is somehow complex I decided to split
up the two things.
This patch makes it easier to paint the overlays for bridges and tunnels
(it reduces the code), so I'll post a new patch for this soon.
The goal of this patch is not totally new, I replaced some code which
tried to paint turns nicer and smoother but didn't cover everything. (for
everyone interested, just have a look at the attached screenshot: the
waysegments are not rectangle) .
I'd like to thank sk750 for helping me to test this patch as he found some
strange artifacts in painting while routing which I didn't see in my
emulator. (...) The patch only connects all waysegments
within a way, so every way will still have the old look at its end. Maybe I
can change this when looking at the crossings of ways ;-)

As I replaced some existing lines of code, the performance seems to be
good. I compared painting the map 50times with and without the patch seeing
only 450ms difference which is less than 1%. I will do some further testing
as I'd like to improve some other parts of ways.java, but this should be
done in a seperate patch."

Thx to pos_ei_don_ for this great patch!