CellID: Include a method to retrieve cell-ids for Motorola phones
There are still a couple of problems with this.
For one, it uses the imsi to get at the mcc and mnc. This will presumably
break when roaming in a different network, as the imsi doesn't change when
roaming. Also imsi seems to have different formats in different countries, so
unless we special case different countries we might get the mnc wrong.
Secondly in the tests reported by Pavel, the mnc and mcc seem swapped, but I can't
quite see where that is coming from.

I am committing this never the less to help testing and debugging, but please do not
use this to upload data to OpenCellID.org, as it may introduce corrupt data!

Thanks to Pavel for pointing out the System.properties for Motorola and helping testing.